Refector network and admin sql
